@@ -5,62 +5,63 @@
   "scripts": {
     "start": "cross-env SOCKET=true node ./scripts/next-remote-watch.js ./data",
     "dev": "next dev",
-    "build": "next build && node ./scripts/generate-sitemap",
+    "build": "next build && node ./scripts/generate-sitemap.mjs",
     "serve": "next start",
     "analyze": "cross-env ANALYZE=true next build",
     "lint": "next lint --fix --dir pages --dir components --dir lib --dir layouts --dir scripts",
     "prepare": "husky install"
   },
   "dependencies": {
-    "@fontsource/inter": "4.5.2",
-    "@mailchimp/mailchimp_marketing": "^3.0.58",
-    "@tailwindcss/forms": "^0.4.0",
-    "@tailwindcss/typography": "^0.5.0",
-    "autoprefixer": "^10.4.0",
-    "esbuild": "^0.13.13",
+    "@fontsource/inter": "^4.5.11",
+    "@mailchimp/mailchimp_marketing": "^3.0.78",
+    "@tailwindcss/forms": "^0.5.2",
+    "@tailwindcss/typography": "^0.5.4",
+    "autoprefixer": "^10.4.8",
+    "esbuild": "^0.14.53",
     "github-slugger": "^1.3.0",
     "gray-matter": "^4.0.2",
-    "image-size": "1.0.0",
-    "mdx-bundler": "^8.0.0",
-    "next": "12.1.4",
+    "image-size": "^1.0.2",
+    "mdx-bundler": "^8.1.0",
+    "next": "^12.2.4",
+    "next-plugin-preact": "^3.0.7",
     "next-themes": "^0.0.14",
     "postcss": "^8.4.5",
-    "preact": "^10.6.2",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "reading-time": "1.3.0",
-    "rehype-autolink-headings": "^6.1.0",
+    "preact": "^10.10.1",
+    "react": "npm:@preact/compat@^17.1.1",
+    "react-dom": "npm:@preact/compat@^17.1.1",
+    "reading-time": "^1.5.0",
+    "rehype-autolink-headings": "^6.1.1",
     "rehype-citation": "^0.4.0",
     "rehype-katex": "^6.0.2",
     "rehype-preset-minify": "6.0.0",
-    "rehype-prism-plus": "^1.1.3",
-    "rehype-slug": "^5.0.0",
+    "rehype-prism-plus": "^1.4.2",
+    "rehype-slug": "^5.0.1",
     "remark-footnotes": "^4.0.1",
     "remark-gfm": "^3.0.1",
     "remark-math": "^5.1.1",
-    "sharp": "^0.28.3",
-    "tailwindcss": "^3.0.23",
+    "sharp": "^0.30.7",
+    "tailwindcss": "^3.1.8",
     "unist-util-visit": "^4.0.0"
   },
   "devDependencies": {
-    "@next/bundle-analyzer": "12.1.4",
-    "@svgr/webpack": "^6.1.2",
+    "@next/bundle-analyzer": "^12.2.4",
+    "@svgr/webpack": "^6.3.1",
     "cross-env": "^7.0.3",
     "dedent": "^0.7.0",
-    "eslint": "^7.29.0",
-    "eslint-config-next": "12.1.4",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^3.3.1",
+    "eslint": "^7.32.0",
+    "eslint-config-next": "^12.2.4",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.2.1",
     "file-loader": "^6.0.0",
-    "globby": "11.0.3",
+    "globby": "^13.1.2",
     "husky": "^6.0.0",
-    "inquirer": "^8.1.1",
-    "lint-staged": "^11.0.0",
+    "inquirer": "^8.2.4",
+    "lint-staged": "^13.0.3",
     "next-remote-watch": "^1.0.0",
-    "prettier": "^2.5.1",
-    "prettier-plugin-tailwindcss": "^0.1.4",
-    "socket.io": "^4.4.0",
-    "socket.io-client": "^4.4.0"
+    "prettier": "2.7",
+    "prettier-plugin-tailwindcss": "^0.1.13",
+    "socket.io": "^4.5.1",
+    "socket.io-client": "^4.5.1"
   },
   "lint-staged": {
     "*.+(js|jsx|ts|tsx)": [
