@@ -1,8 +1,8 @@
-const fs = require('fs')
-const globby = require('globby')
-const matter = require('gray-matter')
-const prettier = require('prettier')
-const siteMetadata = require('../data/siteMetadata')
+import * as fs from 'fs';
+import { globby } from 'globby';
+import matter  from 'gray-matter';
+import prettier   from 'prettier';
+import siteMetadata from '../data/siteMetadata.js';
 
 ;(async () => {
   const prettierConfig = await prettier.resolveConfig('./.prettierrc.js')
